+++ tests/e2e/cli/test_cli_api_equivalence.py
@@ -211,42 +211,3 @@ class TestCLIExpEquivalence:
             cli_pareto_df = pd.read_csv(pareto_path)
             api_pareto_df = api_result.pareto_fronts[exp_name]
             _assert_dataframes_equal(api_pareto_df, cli_pareto_df, f"{exp_name}/pareto_fronts")
-
-
-class TestCLISupportEquivalence:
-    """Tests that cli_support API produces same results as CLI command."""
-
-    def test_cli_support_api_vs_command(self):
-        """cli_support API should return same support status as CLI command."""
-        from aiconfigurator.cli import cli_support
-
-        # Run via Python API
-        api_result = cli_support("Qwen/Qwen3-32B", "h200_sxm")
-
-        # Run via CLI command
-        cmd = [
-            sys.executable,
-            "-m",
-            "aiconfigurator.main",
-            "cli",
-            "support",
-            "--model_path",
-            "Qwen/Qwen3-32B",
-            "--system",
-            "h200_sxm",
-        ]
-
-        result = subprocess.run(cmd, capture_output=True, text=True, timeout=30)
-        assert result.returncode == 0, f"CLI failed: {result.stderr}"
-
-        # Parse CLI output for support status
-        cli_agg_supported = "Aggregated Support:    YES" in result.stdout
-        cli_disagg_supported = "Disaggregated Support: YES" in result.stdout
-
-        # Compare results
-        assert api_result.agg_supported == cli_agg_supported, (
-            f"Aggregated support mismatch: API={api_result.agg_supported}, CLI={cli_agg_supported}"
-        )
-        assert api_result.disagg_supported == cli_disagg_supported, (
-            f"Disaggregated support mismatch: API={api_result.disagg_supported}, CLI={cli_disagg_supported}"
-        )
